requirements.txt
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
scikit-learn>=1.3.0
openvino>=2024.0.0
absl-py>=2.0.0
mediapipe>=0.10.0
